Address pinned installer review feedback (#271)
Centralize checked Git commands and separate workspace progress reporting
from checkout mutation. Split real-Git and pinned-ref test support into
focused modules so touched files remain below the 400-line limit.

Parameterize duplicate prebuilt success cases, share BDD output handling,
and restore test and scenario attributes lost during the rebase. Repair the
branch documentation so spelling, Markdown, and helper ownership contracts
pass repository gates.
